restructure ml to folder machine learning
